v1.3
Resolved an issue in the code that was causing problems when duplicating files. Introduced options to transfer (copy) files from the backup save directory to the main save directory. Additionally, added a feature to display more information. Currently, it shows the last time the save files were accessed, with plans to expand functionality in the future.